+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Changed
 - Boolean processing options such as `enlarge` and `extend` are properly parsed. `1`, `t`, `TRUE`, `true`, `True` are truthy, `0`, `f`, `F`, `FALSE`, `false`, `False` are falsy. All other values are treated as falsy and generate a warning message.
 
+### Fixed
+- Fix segfaults on watermarking in somme cases
+
 ## [2.6.2] - 2019-11-11
 ### Fixed
 - Fix `format` option in presets.